[FLINK-12070] [network] (part 4) Move pipelined subpartition specific code into PipelinedSubpartition class.
